requirements.txt
apache-beam[gcp]
oauth2client==3.0.0
google-cloud-vision
google-cloud-storage
google-cloud-pubsub